leshan: refactor file directory
Rename com.example to lwm2m to simplify the folder chierarchy and make
logs shorter.
